Restart daemon on wallet encrypt
